Integer extraction no longer funnels through i64 — every target throughu64 narrows exactly from the i128 impl, closing the false rejection of the (i64::MAX, u64::MAX] band. Propagated across de/query/money, the formatter, LSP, and validate

+++ nml-cli/examples/serde_deserialize.rs
@@ -9,7 +9,10 @@ use serde::Deserialize;
 #[derive(Deserialize, Debug)]
 struct ServiceConfig {
     host: String,
-    port: f64,
+    // Integer fields deserialize EXACTLY (RFC 0016): out-of-range or
+    // fractional values error — nothing rounds. f64 is only for values
+    // that are genuinely binary floating point.
+    port: u16,
     debug: bool,
     tags: Vec<String>,
 }
